ActionNodes don't reload the main script per run
Previously, ActionNodes with lua scripts would reload
the script containing the action function before running it.

Now, ActionNodes loads the lua script once, places the action
function in a new global with a name prefixed with
"BT_ActionNodeActionFunction_"
and postfixed with a number unique to the ActionNode, and
will execute the new global on activate().

The problem with this method of "caching" the action function is
that it cannot be overwritten (without prior knowledge of the global's
name) during execution.