Implemented CustomNodes
CustomNodes are nodes with logic based on a lua function.
Functions "getChildrenSize" and "activateChild" are available
to the lua function.
CustomLuaNodes use a separate lua state different from the lua state
in BehaviorLuaFactory.
The lua state in CustomLuaNodes uses the extraspace (lua_getextraspace)
so that the custom scripts can use "getChildrenSize" and "activateChild".
Thus, a void pointer is provided in CustomLuaNodeState that can be used
to point to anything.