Implemented assigning IDs to generated nodes
When generating a tree with a BehaviorLuaFactory, an "id" field
can be assigned to a node. It must be a string. Assigning ids allows
for finding a specific node in the tree:
BehaviorNode* ptr = rootNode->findByID("actionNode2");
Note this returns null if there is no node in the tree with that id.